Refactored POTextStyle
