+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/field/POFieldStateStyle.kt
@@ -3,7 +3,7 @@ package com.processout.sdk.ui.core.style.field
 import android.os.Parcelable
 import androidx.annotation.ColorRes
 import com.processout.sdk.ui.core.style.POBorderStyle
-import com.processout.sdk.ui.core.style.text.POTextStyle
+import com.processout.sdk.ui.core.style.POTextStyle
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
