+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/radio/PORadioButtonStyle.kt
@@ -2,7 +2,7 @@ package com.processout.sdk.ui.core.style.radio
 
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
-import com.processout.sdk.ui.core.style.text.POTextStyle
+import com.processout.sdk.ui.core.style.POTextStyle
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
